nvme: move submit_io to src/io-cmds
This is to move other IO commands as same.
